refactor: 리스너를 한번만 등록하도록 변경
