+++ src/main/java/com/quickmemo/plugin/action/OpenMemoAction.java
@@ -5,12 +5,8 @@ import com.intellij.openapi.actionSystem.AnActionEvent;
 import com.intellij.openapi.project.Project;
 import com.intellij.openapi.wm.ToolWindow;
 import com.intellij.openapi.wm.ToolWindowManager;
-import com.intellij.openapi.wm.ex.ToolWindowManagerListener;
-import com.quickmemo.plugin.window.component.MemoEditor;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
-
-import javax.swing.*;
 
 public class OpenMemoAction extends AnAction {
 
@@ -32,32 +28,15 @@ public class OpenMemoAction extends AnAction {
             return;
         }
 
+        toggleWindow(window);
+    }
+
+    private void toggleWindow(ToolWindow window) {
         if (window.isVisible()) {
             window.hide();
             return;
         }
-
-        project.getMessageBus().connect().subscribe(
-                ToolWindowManagerListener.TOPIC,
-                getWindowManagerListener()
-        );
-        window.show(null);
-    }
-
-    private ToolWindowManagerListener getWindowManagerListener() {
-        return new ToolWindowManagerListener() {
-            @Override
-            public void toolWindowShown(@NotNull ToolWindow shownWindow) {
-                if (!WINDOW_NAME.equals(shownWindow.getId())) {
-                    return;
-                }
-
-                SwingUtilities.invokeLater(() -> {
-                    MemoEditor memoEditor = MemoEditor.getInstance();
-                    memoEditor.requestFocusOnEditor();
-                });
-            }
-        };
+        window.show();
     }
 
     private @Nullable ToolWindow getWindow(Project project) {
