+++ src/main/java/com/quickmemo/plugin/window/MemoToolWindowFactory.java
@@ -4,8 +4,10 @@ import com.intellij.openapi.project.Project;
 import com.intellij.openapi.wm.ToolWindow;
 import com.intellij.openapi.wm.ToolWindowFactory;
 import com.intellij.openapi.wm.ToolWindowType;
+import com.intellij.openapi.wm.ex.ToolWindowManagerListener;
 import com.intellij.ui.content.Content;
 import com.intellij.ui.content.ContentFactory;
+import com.quickmemo.plugin.window.component.MemoEditor;
 import org.jetbrains.annotations.NotNull;
 
 import javax.swing.*;
@@ -14,6 +16,7 @@ import java.awt.*;
 public class MemoToolWindowFactory implements ToolWindowFactory {
     private static final int DEFAULT_SIZE = 400;
     private static final Dimension DEFAULT_DIMENSION_SIZE = new Dimension(DEFAULT_SIZE, DEFAULT_SIZE);
+    public static final String WINDOW_NAME = "QuickMemo";
 
     @Override
     public void createToolWindowContent(@NotNull Project project, @NotNull ToolWindow toolWindow) {
@@ -29,6 +32,8 @@ public class MemoToolWindowFactory implements ToolWindowFactory {
                 .createContent(component, "", false);
         toolWindow.getContentManager().addContent(content);
 
+        addEditorOpenListener(project);
+
         SwingUtilities.invokeLater(() -> {
             Timer timer = new Timer(100, event -> {
                 Window win = SwingUtilities.getWindowAncestor(component);
@@ -42,5 +47,28 @@ public class MemoToolWindowFactory implements ToolWindowFactory {
             timer.setRepeats(false);
             timer.start();
         });
+    }
+
+    private void addEditorOpenListener(@NotNull Project project) {
+        project.getMessageBus().connect().subscribe(
+                ToolWindowManagerListener.TOPIC,
+                getWindowManagerListener()
+        );
+    }
+
+    private ToolWindowManagerListener getWindowManagerListener() {
+        return new ToolWindowManagerListener() {
+            @Override
+            public void toolWindowShown(@NotNull ToolWindow shownWindow) {
+                if (!WINDOW_NAME.equals(shownWindow.getId())) {
+                    return;
+                }
+
+                SwingUtilities.invokeLater(() -> {
+                    MemoEditor memoEditor = MemoEditor.getInstance();
+                    memoEditor.requestFocusOnEditor();
+                });
+            }
+        };
     }
 }
